requirements.txt
colorama
dnspython